refactor(banking): extract the broken-bank definitions into a shared concern
Both notify commands carry one half of what "this bank is broken" means, and
a report on the same question needs both. Moving the predicates into
QueriesBankFailures keeps one definition instead of a second, subtly different
copy, and leaves the commands' behaviour untouched.
